+++ extension/oteleventually/eventstore.go
@@ -28,10 +28,10 @@ type InstrumentedEventStore struct {
 	eventStore event.Store
 
 	appendCount    metric.Int64Counter
-	appendDuration metric.Int64ValueRecorder
+	appendDuration metric.Int64Histogram
 
 	streamCount    metric.Int64Counter
-	streamDuration metric.Int64ValueRecorder
+	streamDuration metric.Int64Histogram
 }
 
 func (es *InstrumentedEventStore) registerMetrics() error {
@@ -44,7 +44,7 @@ func (es *InstrumentedEventStore) registerMetrics() error {
 		return fmt.Errorf("oteleventually: failed to register metric: %w", err)
 	}
 
-	if es.appendDuration, err = es.meter.NewInt64ValueRecorder(
+	if es.appendDuration, err = es.meter.NewInt64Histogram(
 		"eventually.events.append.duration.milliseconds",
 		metric.WithUnit(unit.Milliseconds),
 		metric.WithDescription("Duration in milliseconds of append operations performed"),
@@ -59,7 +59,7 @@ func (es *InstrumentedEventStore) registerMetrics() error {
 		return fmt.Errorf("oteleventually: failed to register metric: %w", err)
 	}
 
-	if es.streamDuration, err = es.meter.NewInt64ValueRecorder(
+	if es.streamDuration, err = es.meter.NewInt64Histogram(
 		"eventually.events.stream.duration.milliseconds",
 		metric.WithUnit(unit.Milliseconds),
 		metric.WithDescription("Duration in milliseconds of stream operations performed"),